added time correlation

+++ include/scatterdata.hpp
@@ -28,19 +28,19 @@
 // other headers
 #include "coor3d.hpp"
 
-class ScatterdataKey : public std::pair<CartesianCoor3D,int> {
+class ScatterdataKey : public std::pair<CartesianCoor3D,size_t> {
 	// make this class serializable to 
 	// allow sample to be transmitted via MPI
     friend class boost::serialization::access;	
 	template<class Archive> void serialize(Archive & ar, const unsigned int version)
     {
-		ar & boost::serialization::base_object<std::pair<CartesianCoor3D,int> >(*this);
+		ar & boost::serialization::base_object<std::pair<CartesianCoor3D,size_t> >(*this);
     }
 	///////////////////
 	
 public:
 	ScatterdataKey() {}
-	ScatterdataKey(CartesianCoor3D q, int frame) { this->first = q; this->second = frame; }	
+	ScatterdataKey(CartesianCoor3D q, size_t frame) { this->first = q; this->second = frame; }	
 	
 	bool operator <(const ScatterdataKey that) const;
 };
